Wrapper functions for the POT library

If you load the PyCall package, you get access to wrappers of functions in the Python Optimal Transport (POT) package. The wrapper functions are available in the submodule POT.

Exact optimal transport (Kantorovich) problem

OptimalTransport.POT.emd — Function

emd(mu, nu, C)

Compute transport map for Monge-Kantorovich problem with source and target marginals mu and nu and a cost matrix C of dimensions (length(mu), length(nu)).

Return optimal transport coupling γ of the same dimensions as C which solves

\[\inf_{\gamma \in \Pi(\mu, \nu)} \langle \gamma, C \rangle\]

This function is a wrapper of the function emd in the Python Optimal Transport package.

OptimalTransport.POT.emd2 — Function

emd2(mu, nu, C)

Compute exact transport cost for Monge-Kantorovich problem with source and target marginals mu and nu and a cost matrix C of dimensions (length(mu), length(nu)).

Returns optimal transport cost (a scalar), i.e. the optimal value

\[\inf_{\gamma \in \Pi(\mu, \nu)} \langle \gamma, C \rangle\]

This function is a wrapper of the function emd2 in the Python Optimal Transport package.

Entropically regularised optimal transport

OptimalTransport.POT.sinkhorn — Function

sinkhorn(mu, nu, C, eps; tol=1e-9, max_iter = 1000, method = "sinkhorn", verbose = false)

Compute optimal transport map of histograms mu and nu with cost matrix C and entropic regularization parameter eps.

Method can be a choice of "sinkhorn", "greenkhorn", "sinkhorn_stabilized", or "sinkhorn_epsilon_scaling" (Flamary et al., 2017).

This function is a wrapper of the function sinkhorn in the Python Optimal Transport package.

OptimalTransport.POT.sinkhorn2 — Function

sinkhorn2(mu, nu, C, eps; tol=1e-9, max_iter = 1000, method = "sinkhorn", verbose = false)

Compute optimal transport cost of histograms mu and nu with cost matrix C and entropic regularization parameter eps.

Method can be a choice of "sinkhorn", "greenkhorn", "sinkhorn_stabilized", or "sinkhorn_epsilon_scaling" (Flamary et al., 2017).

This function is a wrapper of the function sinkhorn2 in the Python Optimal Transport package.

Unbalanced optimal transport

OptimalTransport.POT.sinkhorn_unbalanced — Function

sinkhorn_unbalanced(mu, nu, C, eps, lambda; tol = 1e-9, max_iter = 1000, method = "sinkhorn", verbose = false)

Compute optimal transport map of histograms mu and nu with cost matrix C, using entropic regularisation parameter eps and marginal weighting functions lambda.

This function is a wrapper of the function sinkhorn_unbalanced in the Python Optimal Transport package.

OptimalTransport.POT.sinkhorn_unbalanced2 — Function

sinkhorn_unbalanced2(mu, nu, C, eps, lambda; tol = 1e-9, max_iter = 1000, method = "sinkhorn", verbose = false)

Compute optimal transport cost of histograms mu and nu with cost matrix C, using entropic regularisation parameter eps and marginal weighting functions lambda.

This function is a wrapper of the function sinkhorn_unbalanced2 in the Python Optimal Transport package.