OptimalTransport.jl Documentation

Exact optimal transport (Kantorovich) problem

OptimalTransport.emd — Function

emd(μ, ν, C, optimizer)

Compute the optimal transport plan γ for the Monge-Kantorovich problem with source histogram μ, target histogram ν, and cost matrix C of size (length(μ), length(ν)) which solves

\[\inf_{γ ∈ Π(μ, ν)} \langle γ, C \rangle.\]

The corresponding linear programming problem is solved with the user-provided optimizer. Possible choices are Tulip.Optimizer() and Clp.Optimizer() in the Tulip and Clp packages, respectively.

OptimalTransport.emd2 — Function

emd2(μ, ν, C, optimizer; plan=nothing)

Compute the optimal transport cost (a scalar) for the Monge-Kantorovich problem with source histogram μ, target histogram ν, and cost matrix C of size (length(μ), length(ν)) which is given by

\[\inf_{γ ∈ Π(μ, ν)} \langle γ, C \rangle.\]

The corresponding linear programming problem is solved with the user-provided optimizer. Possible choices are Tulip.Optimizer() and Clp.Optimizer() in the Tulip and Clp packages, respectively.

A pre-computed optimal transport plan may be provided.

Entropically regularised optimal transport

OptimalTransport.sinkhorn — Function

sinkhorn(mu, nu, C, eps; tol=1e-9, check_marginal_step=10, maxiter=1000)

Compute entropically regularised transport plan of histograms mu and nu with cost matrix C and entropic regularization parameter eps.

Return optimal transport coupling γ of the same dimensions as C which solves

\[\inf_{\gamma \in \Pi(\mu, \nu)} \langle \gamma, C \rangle - \epsilon H(\gamma)\]

where $H$ is the entropic regulariser, $H(\gamma) = -\sum_{i, j} \gamma_{ij} \log(\gamma_{ij})$.

OptimalTransport.sinkhorn2 — Function

sinkhorn2(mu, nu, C, eps; plan=nothing, kwargs...)

Compute entropically regularised transport cost of histograms mu and nu with cost matrix C and entropic regularization parameter eps.

Return optimal value of

\[\inf_{\gamma \in \Pi(\mu, \nu)} \langle \gamma, C \rangle - \epsilon H(\gamma)\]

where $H$ is the entropic regulariser, $H(\gamma) = -\sum_{i, j} \gamma_{ij} \log(\gamma_{ij})$.

A pre-computed optimal transport plan may be provided.

See also: sinkhorn

OptimalTransport.sinkhorn_stabilized_epsscaling — Function

sinkhorn_stabilized_epsscaling(mu, nu, C, eps; absorb_tol = 1e3, max_iter = 1000, tol = 1e-9, lambda = 0.5, k = 5, verbose = false)

Compute optimal transport plan of histograms mu and nu with cost matrix C and entropic regularisation parameter eps. Uses stabilized Sinkhorn algorithm with epsilon-scaling (Schmitzer et al., 2019).

k epsilon-scaling steps are used with scaling factor lambda, i.e. sequentially solve Sinkhorn with regularisation parameters [lambda^(1-k), ..., lambda^(-1), 1]*eps.

OptimalTransport.sinkhorn_stabilized — Function

sinkhorn_stabilized(mu, nu, C, eps; absorb_tol = 1e3, max_iter = 1000, tol = 1e-9, alpha = nothing, beta = nothing, return_duals = false, verbose = false)

Compute optimal transport plan of histograms mu and nu with cost matrix C and entropic regularisation parameter eps. Uses stabilized Sinkhorn algorithm (Schmitzer et al., 2019).

OptimalTransport.sinkhorn_barycenter — Function

sinkhorn_barycenter(mu_all, C_all, eps, lambda_all; tol = 1e-9, check_marginal_step = 10, max_iter = 1000)

Compute the entropically regularised (i.e. Sinkhorn) barycenter for a collection of N histograms mu_all with respective cost matrices C_all, relative weights lambda_all, and entropic regularisation parameter eps.

mu_all is taken to contain N histograms mu_all[i, :] for math i = 1, \ldots, N.

C_all is taken to be a list of N cost matrices corresponding to the mu_all[i, :].

eps is the scalar regularisation parameter.

lambda_all are positive weights.

Returns the entropically regularised barycenter of the mu_all, i.e. the distribution that minimises

\[\min_{\mu \in \Sigma} \sum_{i = 1}^N \lambda_i \mathrm{entropicOT}^{\epsilon}_{C_i}(\mu, \mu_i)\]

where $\mathrm{entropicOT}^{\epsilon}_{C}$ denotes the entropic optimal transport cost with cost $C$ and entropic regularisation level $\epsilon$.

Unbalanced optimal transport

OptimalTransport.sinkhorn_unbalanced — Function

sinkhorn_unbalanced(mu, nu, C, lambda1, lambda2, eps; tol = 1e-9, max_iter = 1000, verbose = false, proxdiv_F1 = nothing, proxdiv_F2 = nothing)

Computes the optimal transport plan of histograms mu and nu with cost matrix C and entropic regularization parameter eps, using the unbalanced Sinkhorn algorithm [Chizat 2016] with KL-divergence terms for soft marginal constraints, with weights (lambda1, lambda2) for the marginals mu, nu respectively.

For full generality, the user can specify the soft marginal constraints $(F_1(\cdot | \mu), F_2(\cdot | \nu))$ to the problem

\[\min_\gamma \epsilon \mathrm{KL}(\gamma | \exp(-C/\epsilon)) + F_1(\gamma_1 | \mu) + F_2(\gamma_2 | \nu)\]

via math\mathrm{proxdiv}_{F_1}(s, p) and math\mathrm{proxdiv}_{F_2}(s, p) (see Chizat et al., 2016 for details on this). If specified, the algorithm will use the user-specified F1, F2 rather than the default (a KL-divergence).

OptimalTransport.sinkhorn_unbalanced2 — Function

sinkhorn_unbalanced2(mu, nu, C, lambda1, lambda2, eps; plan=nothing, kwargs...)

Computes the optimal transport cost of histograms mu and nu with cost matrix C and entropic regularization parameter eps, using the unbalanced Sinkhorn algorithm [Chizat 2016] with KL-divergence terms for soft marginal constraints, with weights (lambda1, lambda2) for the marginals mu, nu respectively.

A pre-computed optimal transport plan may be provided.

See also: sinkhorn_unbalanced

Quadratically regularised optimal transport

OptimalTransport.quadreg — Function

quadreg(mu, nu, C, ϵ; θ = 0.1, tol = 1e-5,maxiter = 50,κ = 0.5,δ = 1e-5)

Computes the optimal transport plan of histograms mu and nu with cost matrix C and quadratic regularization parameter ϵ, using the semismooth Newton algorithm [Lorenz 2016].

This implementation makes use of IterativeSolvers.jl and SparseArrays.jl.

Parameters:

θ: starting Armijo parameter.

tol: tolerance of marginal error.

maxiter: maximum interation number.

κ: control parameter of Armijo.

δ: small constant for the numerical stability of conjugate gradient iterative solver.

Tips: If the algorithm does not converge, try some different values of θ.

Reference: Lorenz, D.A., Manns, P. and Meyer, C., 2019. Quadratically regularized optimal transport. arXiv preprint arXiv:1903.01112v4.